reaktoro/requirements.txt
flask
reaktoro
waitress
grpcio
grpcio-tools
# other dependencies
